+++ libs/gbd-mapping/docs/source/conf.py
@@ -10,36 +10,23 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-from pathlib import Path
+import datetime
 
 from docutils.nodes import Text
 from sphinx.ext.intersphinx import missing_reference
 
-import gbd_mapping
-
-base_dir = Path(gbd_mapping.__file__).parent
-
-about = {}
-with (base_dir / "__about__.py").open() as f:
-    exec(f.read(), about)
-
-sys.path.insert(0, str(Path("..").resolve()))
+import vivarium.gbd_mapping
 
 # -- Project information -----------------------------------------------------
 
-project = about["__title__"]
-copyright = f'2023, {about["__author__"]}'
-author = about["__author__"]
+project = "vivarium.gbd_mapping"
+author = "The vivarium developers"
+copyright = f"2023-{datetime.date.today().year}, Institute for Health Metrics and Evaluation"
 
 # The short X.Y version.
-version = gbd_mapping.__version__
+version = vivarium.gbd_mapping.__version__
 # The full version, including alpha/beta/rc tags.
-release = gbd_mapping.__version__
+release = vivarium.gbd_mapping.__version__
 
 
 # -- General configuration ------------------------------------------------
@@ -132,7 +119,7 @@ html_sidebars = {
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = f'{about["__title__"]}doc'
+htmlhelp_basename = f"{project}doc"
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -158,9 +145,9 @@ latex_elements = {
 latex_documents = [
     (
         master_doc,
-        f'{about["__title__"]}.tex',
-        f'{about["__title__"]} Documentation',
-        about["__author__"],
+        f"{project}.tex",
+        f"{project} Documentation",
+        author,
         "manual",
     ),
 ]
@@ -170,9 +157,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, f'{about["__title__"]}', f'{about["__title__"]} Documentation', [author], 1)
-]
+man_pages = [(master_doc, project, f"{project} Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output -------------------------------------------
@@ -183,11 +168,11 @@ man_pages = [
 texinfo_documents = [
     (
         master_doc,
-        f'{about["__title__"]}',
-        f'{about["__title__"]} Documentation',
+        project,
+        f"{project} Documentation",
         author,
-        f'{about["__title__"]}',
-        about["__summary__"],
+        project,
+        "A programmatically accessible mapping of GBD entities.",
         "Miscellaneous",
     ),
 ]